+++ agent_debugger_sdk/adapters/tests/langchain/__init__.py
@@ -1,0 +1,8 @@
+"""Tests for LangChain adapter.
+
+This package contains tests for the LangChain adapter, split into focused modules:
+- fixtures: Mock classes for testing
+- test_handler: Tests for LangChainTracingHandler
+- test_adapter: Tests for LangChainAdapter
+- test_errors: Tests for error boundary handling
+"""
